refactor(executor): modularize order.rs ORDER BY handling
Split the 2,164-line order.rs into a modular order/ directory with focused files:

- position.rs (235 lines): Column position extraction, validation, wildcard expansion
- comparison.rs (105 lines): Value comparison with collation support
- parallel.rs (47 lines): Parallel sorting implementation (feature-gated)
- resolution.rs (1,456 lines): Alias resolution for ORDER BY and WHERE clauses
- mod.rs (260 lines): Main API, orchestration, and re-exports

Benefits:
- Smaller, focused files for easier maintenance
- Parallel sorting logic isolated behind feature gate
- Position handling separated from sorting
- Clear separation of concerns

Closes #4848
