+++ crates/vibesql-executor/src/select/order/resolution.rs
@@ -1,409 +1,23 @@
-//! ORDER BY sorting logic
+//! Alias and expression resolution for ORDER BY and WHERE clauses.
+//!
+//! This module handles:
+//! - Resolving ORDER BY column positions and aliases
+//! - Resolving ORDER BY expressions for aggregate queries
+//! - Resolving SELECT aliases in WHERE clauses (SQLite extension)
+//! - Extracting aggregate expressions from ORDER BY for pre-computation
 
-use std::{borrow::Cow, cmp::Ordering};
+use std::borrow::Cow;
 
-#[cfg(feature = "parallel")]
-use rayon::slice::ParallelSliceMut;
-
-use super::grouping::compare_sql_values;
-#[cfg(feature = "parallel")]
-use super::parallel::ParallelConfig;
-use crate::{
-    errors::ExecutorError, evaluator::CombinedExpressionEvaluator, schema::CombinedSchema,
+use super::position::{
+    count_select_columns, extract_column_position, resolve_position_to_column_name,
+    resolve_position_with_wildcards, validate_column_position, ColumnPositionResult,
+    ResolvedPosition,
 };
+use crate::{errors::ExecutorError, schema::CombinedSchema};
 
-/// Result of extracting a column position from an ORDER BY expression
-#[derive(Debug, Clone, Copy)]
-enum ColumnPositionResult {
-    /// A valid positive column position (1-indexed)
-    Position(i64),
-    /// A negative column position (always invalid, but we track it for error reporting)
-    Negative(i64),
-    /// Not a column position expression
-    NotAPosition,
-}
-
-/// Extracts a numeric column position from an ORDER BY expression.
-///
-/// Handles:
-/// - `ORDER BY N` - Integer literal → Position(N)
-/// - `ORDER BY +N` - Unary plus with integer → Position(N) (#4418)
-/// - `ORDER BY -N` - Unary minus with integer → Negative(N)
-/// - Other expressions → NotAPosition
-fn extract_column_position(expr: &vibesql_ast::Expression) -> ColumnPositionResult {
-    match expr {
-        // Direct integer literal: ORDER BY 1, ORDER BY 2, etc.
-        vibesql_ast::Expression::Literal(vibesql_types::SqlValue::Integer(pos)) => {
-            ColumnPositionResult::Position(*pos)
-        }
-        // Unary operator with integer
-        vibesql_ast::Expression::UnaryOp { op, expr } => {
-            if let vibesql_ast::Expression::Literal(vibesql_types::SqlValue::Integer(pos)) =
-                expr.as_ref()
-            {
-                match op {
-                    // ORDER BY +N: treat as ORDER BY N (#4418)
-                    vibesql_ast::UnaryOperator::Plus => ColumnPositionResult::Position(*pos),
-                    // ORDER BY -N: always invalid
-                    vibesql_ast::UnaryOperator::Minus => ColumnPositionResult::Negative(*pos),
-                    _ => ColumnPositionResult::NotAPosition,
-                }
-            } else {
-                ColumnPositionResult::NotAPosition
-            }
-        }
-        _ => ColumnPositionResult::NotAPosition,
-    }
-}
-
-/// Validates a column position and returns an error if out of range.
-fn validate_column_position(
-    pos: i64,
-    column_count: usize,
-    term_index: usize,
-) -> Result<usize, ExecutorError> {
-    if pos <= 0 || (pos as usize) > column_count {
-        return Err(ExecutorError::OrderByOutOfRange {
-            term_position: term_index + 1, // Convert to 1-indexed for error message
-            column_number: pos,
-            select_list_len: column_count,
-        });
-    }
-    Ok((pos as usize) - 1) // Convert to 0-indexed
-}
-
-/// Resolves a column position to a column name for aggregate queries.
-///
-/// For aggregate queries, we simply look up the alias/expression at the given position
-/// in the SELECT list without wildcard expansion (aggregates don't typically use wildcards).
-fn resolve_position_to_column_name(
-    idx: usize,
-    select_list: &[vibesql_ast::SelectItem],
-) -> Option<String> {
-    if let vibesql_ast::SelectItem::Expression { expr, alias, .. } = &select_list[idx] {
-        Some(if let Some(alias_name) = alias {
-            alias_name.clone()
-        } else if let vibesql_ast::Expression::ColumnRef(col_id) = expr {
-            col_id.column_canonical().to_string()
-        } else if let vibesql_ast::Expression::AggregateFunction { name, .. } = expr {
-            // For aggregate functions without alias, use the function name (lowercase)
-            // This matches the schema column name generated by the aggregate executor
-            name.to_lowercase()
-        } else {
-            format!("col{}", idx + 1)
-        })
-    } else {
-        None
-    }
-}
-
-/// Result of resolving a column position with wildcard expansion
-enum ResolvedPosition<'a> {
-    /// The expression at the position (for regular SELECT items)
-    Expression(&'a vibesql_ast::Expression),
-    /// A column name resolved from a wildcard expansion
-    ColumnName(String),
-    /// Position not found (shouldn't happen if validation passed)
-    NotFound,
-}
-
-/// Resolves a column position to an expression or column name, handling wildcard expansion.
-///
-/// This handles SELECT lists that may contain wildcards (`*`) or qualified wildcards (`table.*`)
-/// which expand to multiple columns. The position is 0-indexed.
-fn resolve_position_with_wildcards<'a>(
-    target_col: usize,
-    select_list: &'a [vibesql_ast::SelectItem],
-    schema: Option<&CombinedSchema>,
-) -> ResolvedPosition<'a> {
-    let mut current_col = 0;
-
-    for item in select_list {
-        match item {
-            vibesql_ast::SelectItem::Wildcard { .. } => {
-                // Count how many columns this wildcard expands to
-                let wildcard_cols = if let Some(s) = schema {
-                    s.table_schemas.values().map(|(_, ts)| ts.columns.len()).sum()
-                } else {
-                    1
-                };
-
-                if target_col < current_col + wildcard_cols {
-                    // The position is within this wildcard's expanded columns
-                    if let Some(s) = schema {
-                        let offset_in_wildcard = target_col - current_col;
-                        // Collect all columns from all tables in schema order
-                        let mut all_columns: Vec<(usize, String)> = Vec::new();
-                        for (start_idx, table_schema) in s.table_schemas.values() {
-                            for (i, col) in table_schema.columns.iter().enumerate() {
-                                all_columns.push((start_idx + i, col.name.clone()));
-                            }
-                        }
-                        all_columns.sort_by_key(|(idx, _)| *idx);
-
-                        if offset_in_wildcard < all_columns.len() {
-                            return ResolvedPosition::ColumnName(
-                                all_columns[offset_in_wildcard].1.clone(),
-                            );
-                        }
-                    }
-                    return ResolvedPosition::NotFound;
-                }
-                current_col += wildcard_cols;
-            }
-            vibesql_ast::SelectItem::QualifiedWildcard { qualifier, .. } => {
-                // Count how many columns this qualified wildcard expands to
-                let qualified_cols = if let Some(s) = schema {
-                    s.get_table(qualifier).map(|(_, ts)| ts.columns.len()).unwrap_or(1)
-                } else {
-                    1
-                };
-
-                if target_col < current_col + qualified_cols {
-                    // The position is within this qualified wildcard's columns
-                    if let Some(s) = schema {
-                        if let Some((_, table_schema)) = s.get_table(qualifier) {
-                            let offset = target_col - current_col;
-                            if offset < table_schema.columns.len() {
-                                return ResolvedPosition::ColumnName(
-                                    table_schema.columns[offset].name.clone(),
-                                );
-                            }
-                        }
-                    }
-                    return ResolvedPosition::NotFound;
-                }
-                current_col += qualified_cols;
-            }
-            vibesql_ast::SelectItem::Expression { expr, .. } => {
-                if target_col == current_col {
-                    return ResolvedPosition::Expression(expr);
-                }
-                current_col += 1;
-            }
-        }
-    }
-
-    ResolvedPosition::NotFound
-}
-
-/// Sort key for ORDER BY: (value, direction, nulls_order, collation)
-type SortKey = (
-    vibesql_types::SqlValue,
-    vibesql_ast::OrderDirection,
-    Option<vibesql_ast::NullsOrder>,
-    Option<String>,
-);
-
-/// Row with optional sort keys for ORDER BY
-pub(super) type RowWithSortKeys = (vibesql_storage::Row, Option<Vec<SortKey>>);
-
-/// Count the number of columns in a SELECT list after wildcard expansion
-///
-/// This correctly handles:
-/// - Regular expressions: count as 1 column each
-/// - Wildcard (`SELECT *`): expands to all columns from the schema
-/// - Qualified wildcards (`SELECT table.*`): expands to all columns from that table
-///
-/// # Arguments
-/// * `select_list` - The SELECT list items
-/// * `schema` - Optional schema for wildcard expansion. If None, wildcards count as 1.
-fn count_select_columns(
-    select_list: &[vibesql_ast::SelectItem],
-    schema: Option<&CombinedSchema>,
-) -> usize {
-    let mut count = 0;
-    for item in select_list {
-        match item {
-            vibesql_ast::SelectItem::Wildcard { .. } => {
-                // Expand wildcard to all columns from schema
-                if let Some(s) = schema {
-                    // Count all columns from all tables in the schema
-                    for (_, table_schema) in s.table_schemas.values() {
-                        count += table_schema.columns.len();
-                    }
-                } else {
-                    // No schema available, count as 1
-                    count += 1;
-                }
-            }
-            vibesql_ast::SelectItem::QualifiedWildcard { qualifier, .. } => {
-                // Expand qualified wildcard to all columns from the specified table
-                if let Some(s) = schema {
-                    if let Some((_, table_schema)) = s.get_table(qualifier) {
-                        count += table_schema.columns.len();
-                    } else {
-                        // Table not found, count as 1 (error will be caught elsewhere)
-                        count += 1;
-                    }
-                } else {
-                    // No schema available, count as 1
-                    count += 1;
-                }
-            }
-            vibesql_ast::SelectItem::Expression { .. } => {
-                count += 1;
-            }
-        }
-    }
-    count
-}
-
-/// Apply ORDER BY sorting to rows
-///
-/// Evaluates ORDER BY expressions for each row and sorts them according to the specified
-/// directions (ASC/DESC). Supports multi-column sorting with stable sort behavior.
-///
-/// ORDER BY can reference:
-/// - Columns from the FROM clause
-/// - Aliases from the SELECT list
-/// - Original column names that have been aliased in SELECT
-/// - Arbitrary expressions
-pub(super) fn apply_order_by(
-    mut rows: Vec<RowWithSortKeys>,
-    order_by: &[vibesql_ast::OrderByItem],
-    evaluator: &CombinedExpressionEvaluator,
-    select_list: &[vibesql_ast::SelectItem],
-) -> Result<Vec<RowWithSortKeys>, ExecutorError> {
-    // Get schema for proper wildcard expansion when counting columns (#4413)
-    let schema = evaluator.schema();
-
-    // Evaluate ORDER BY expressions for each row
-    for (row, sort_keys) in &mut rows {
-        // Clear CSE cache before evaluating this row's ORDER BY expressions
-        // to prevent stale cached column values from previous rows
-        evaluator.clear_cse_cache();
-
-        let mut keys = Vec::new();
-        for (term_index, order_item) in order_by.iter().enumerate() {
-            // Check if ORDER BY expression is a SELECT list alias or matches an aliased column
-            // Evaluator handles window functions via window_mapping if present
-            // Pass schema for proper wildcard expansion in column count validation
-            let expr_to_eval =
-                resolve_order_by_alias(&order_item.expr, select_list, term_index, Some(schema))?;
-            let key_value = evaluator.eval(expr_to_eval.as_ref(), row)?;
-            // Get collation for this ORDER BY expression (explicit or inherited from column)
-            let collation = evaluator.get_expression_collation(expr_to_eval.as_ref());
-            keys.push((key_value, order_item.direction.clone(), order_item.nulls_order, collation));
-        }
-        *sort_keys = Some(keys);
-    }
-
-    // Sort by the evaluated keys (with automatic parallelism based on row count when feature
-    // enabled)
-    let comparison_fn = |(_, keys_a): &RowWithSortKeys, (_, keys_b): &RowWithSortKeys| {
-        let keys_a = keys_a.as_ref().unwrap();
-        let keys_b = keys_b.as_ref().unwrap();
-
-        for ((val_a, dir, nulls_order, collation), (val_b, _, _, _)) in
-            keys_a.iter().zip(keys_b.iter())
-        {
-            // Determine NULL ordering:
-            // - If explicitly specified via NULLS FIRST/LAST, use that
-            // - Default: SQLite treats NULL as smallest value, so:
-            //   - ASC: NULL comes first (smallest first)
-            //   - DESC: NULL comes last (smallest last)
-            let nulls_first = match nulls_order {
-                Some(vibesql_ast::NullsOrder::First) => true,
-                Some(vibesql_ast::NullsOrder::Last) => false,
-                None => matches!(dir, vibesql_ast::OrderDirection::Asc),
-            };
-
-            // Handle NULLs according to nulls_first setting
-            let cmp = match (val_a.is_null(), val_b.is_null()) {
-                (true, true) => Ordering::Equal,
-                (true, false) => {
-                    if nulls_first {
-                        return Ordering::Less; // NULL sorts before non-NULL
-                    } else {
-                        return Ordering::Greater; // NULL sorts after non-NULL
-                    }
-                }
-                (false, true) => {
-                    if nulls_first {
-                        return Ordering::Greater; // non-NULL sorts after NULL
-                    } else {
-                        return Ordering::Less; // non-NULL sorts before NULL
-                    }
-                }
-                (false, false) => {
-                    // Apply collation transformation if needed
-                    let (cmp_val_a, cmp_val_b) = if let Some(ref coll) = collation {
-                        if coll.eq_ignore_ascii_case("nocase") {
-                            // For NOCASE collation, uppercase both string values
-                            let a = match val_a {
-                                vibesql_types::SqlValue::Varchar(s) => {
-                                    Cow::Owned(vibesql_types::SqlValue::Varchar(
-                                        arcstr::ArcStr::from(s.to_uppercase()),
-                                    ))
-                                }
-                                vibesql_types::SqlValue::Character(s) => {
-                                    Cow::Owned(vibesql_types::SqlValue::Character(
-                                        arcstr::ArcStr::from(s.to_uppercase()),
-                                    ))
-                                }
-                                other => Cow::Borrowed(other),
-                            };
-                            let b = match val_b {
-                                vibesql_types::SqlValue::Varchar(s) => {
-                                    Cow::Owned(vibesql_types::SqlValue::Varchar(
-                                        arcstr::ArcStr::from(s.to_uppercase()),
-                                    ))
-                                }
-                                vibesql_types::SqlValue::Character(s) => {
-                                    Cow::Owned(vibesql_types::SqlValue::Character(
-                                        arcstr::ArcStr::from(s.to_uppercase()),
-                                    ))
-                                }
-                                other => Cow::Borrowed(other),
-                            };
-                            (a, b)
-                        } else {
-                            (Cow::Borrowed(val_a), Cow::Borrowed(val_b))
-                        }
-                    } else {
-                        (Cow::Borrowed(val_a), Cow::Borrowed(val_b))
-                    };
-
-                    // Compare non-NULL values, respecting direction
-                    match dir {
-                        vibesql_ast::OrderDirection::Asc => {
-                            compare_sql_values(&cmp_val_a, &cmp_val_b)
-                        }
-                        vibesql_ast::OrderDirection::Desc => {
-                            compare_sql_values(&cmp_val_a, &cmp_val_b).reverse()
-                        }
-                    }
-                }
-            };
-
-            if cmp != Ordering::Equal {
-                return cmp;
-            }
-        }
-        Ordering::Equal
-    };
-
-    #[cfg(feature = "parallel")]
-    {
-        let config = ParallelConfig::global();
-        if config.should_parallelize_sort(rows.len()) {
-            // Parallel sort for large datasets
-            rows.par_sort_by(comparison_fn);
-        } else {
-            // Sequential sort for small datasets
-            rows.sort_by(comparison_fn);
-        }
-    }
-
-    #[cfg(not(feature = "parallel"))]
-    {
-        // Always use sequential sort when parallel feature is disabled
-        rows.sort_by(comparison_fn);
-    }
-
-    Ok(rows)
-}
+// =============================================================================
+// ORDER BY Alias Resolution for Aggregate Queries
+// =============================================================================
 
 /// Resolve ORDER BY expression for aggregate results to result schema column names
 ///
@@ -689,6 +303,10 @@ fn expressions_equal(a: &vibesql_ast::Expression, b: &vibesql_ast::Expression) -
         _ => format!("{:?}", a) == format!("{:?}", b),
     }
 }
+
+// =============================================================================
+// ORDER BY Alias Resolution for Regular Queries
+// =============================================================================
 
 /// Resolve ORDER BY expression that might be a SELECT list alias or column position
 ///
@@ -986,6 +604,10 @@ fn resolve_alias_or_clone(
     resolve_aliases_in_expression(expr, select_list)
 }
 
+// =============================================================================
+// WHERE Clause Alias Resolution (SQLite Extension)
+// =============================================================================
+
 /// Resolve SELECT aliases in WHERE clause expression (SQLite extension).
 ///
 /// SQLite allows referencing SELECT column aliases in WHERE clause, even though
@@ -1046,7 +668,7 @@ pub(crate) fn select_list_has_aliases(select_list: &[vibesql_ast::SelectItem]) -
 pub(crate) fn resolve_where_aliases_with_schema(
     where_expr: &vibesql_ast::Expression,
     select_list: &[vibesql_ast::SelectItem],
-    schema: &crate::schema::CombinedSchema,
+    schema: &CombinedSchema,
 ) -> vibesql_ast::Expression {
     // Fast path: if no SELECT items have aliases, nothing to resolve
     // This is a common case in TPC-C and other OLTP workloads
@@ -1064,7 +686,7 @@ pub(crate) fn resolve_where_aliases_with_schema(
 fn resolve_where_expression_with_schema_ref(
     expr: &vibesql_ast::Expression,
     select_list: &[vibesql_ast::SelectItem],
-    schema: &crate::schema::CombinedSchema,
+    schema: &CombinedSchema,
 ) -> vibesql_ast::Expression {
     use vibesql_ast::Expression;
 
@@ -1681,6 +1303,10 @@ fn resolve_where_expression_with_schema(
     }
 }
 
+// =============================================================================
+// ORDER BY Aggregate Extraction
+// =============================================================================
+
 /// Extract aggregate expressions from ORDER BY items for pre-computation during GROUP BY.
 ///
 /// When ORDER BY contains aggregate functions (e.g., `ORDER BY max(n)+0`), these aggregates
@@ -1826,339 +1452,5 @@ fn collect_aggregates_from_expr(
             collect_aggregates_from_expr(inner, aggregates);
         }
         Expression::MatchAgainst { .. } => {}
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::cmp::Ordering;
-
-    use vibesql_storage::Row;
-    use vibesql_types::SqlValue;
-
-    use super::*;
-
-    /// Test the comparison function logic with pre-evaluated sort keys
-    /// This tests the parallel/sequential sorting logic without needing full evaluator setup
-    #[test]
-    fn test_sort_with_keys_small_dataset() {
-        // Small dataset with pre-populated sort keys
-        let mut rows: Vec<RowWithSortKeys> = vec![
-            (
-                Row::from_vec(vec![SqlValue::Integer(3)]),
-                Some(vec![(SqlValue::Integer(3), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Integer(1)]),
-                Some(vec![(SqlValue::Integer(1), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Integer(2)]),
-                Some(vec![(SqlValue::Integer(2), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-        ];
-
-        // Apply sorting logic (mimics what apply_order_by does after key evaluation)
-        let comparison_fn = |(_, keys_a): &RowWithSortKeys, (_, keys_b): &RowWithSortKeys| {
-            let keys_a = keys_a.as_ref().unwrap();
-            let keys_b = keys_b.as_ref().unwrap();
-
-            for ((val_a, dir, _nulls, _coll), (val_b, _, _, _)) in keys_a.iter().zip(keys_b.iter())
-            {
-                // SQLite treats NULL as smallest value:
-                // - ASC: NULL comes first (Less than non-NULL)
-                // - DESC: NULL comes last (Greater than non-NULL)
-                let cmp = match (val_a.is_null(), val_b.is_null()) {
-                    (true, true) => Ordering::Equal,
-                    (true, false) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Less
-                        } else {
-                            Ordering::Greater
-                        };
-                    }
-                    (false, true) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Greater
-                        } else {
-                            Ordering::Less
-                        };
-                    }
-                    (false, false) => match dir {
-                        vibesql_ast::OrderDirection::Asc => compare_sql_values(val_a, val_b),
-                        vibesql_ast::OrderDirection::Desc => {
-                            compare_sql_values(val_a, val_b).reverse()
-                        }
-                    },
-                };
-
-                if cmp != Ordering::Equal {
-                    return cmp;
-                }
-            }
-            Ordering::Equal
-        };
-
-        #[cfg(feature = "parallel")]
-        {
-            let config = ParallelConfig::global();
-            if config.should_parallelize_sort(rows.len()) {
-                rows.par_sort_by(comparison_fn);
-            } else {
-                rows.sort_by(comparison_fn);
-            }
-        }
-
-        #[cfg(not(feature = "parallel"))]
-        {
-            rows.sort_by(comparison_fn);
-        }
-
-        // Verify sorted order
-        assert_eq!(rows[0].0.values[0], SqlValue::Integer(1));
-        assert_eq!(rows[1].0.values[0], SqlValue::Integer(2));
-        assert_eq!(rows[2].0.values[0], SqlValue::Integer(3));
-    }
-
-    #[test]
-    fn test_sort_with_keys_large_dataset() {
-        // Create large dataset that will trigger parallel path
-        let mut rows: Vec<RowWithSortKeys> = Vec::new();
-        for i in (0..15000).rev() {
-            rows.push((
-                Row::from_vec(vec![SqlValue::Integer(i)]),
-                Some(vec![(SqlValue::Integer(i), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ));
-        }
-
-        let comparison_fn = |(_, keys_a): &RowWithSortKeys, (_, keys_b): &RowWithSortKeys| {
-            let keys_a = keys_a.as_ref().unwrap();
-            let keys_b = keys_b.as_ref().unwrap();
-
-            for ((val_a, dir, _nulls, _coll), (val_b, _, _, _)) in keys_a.iter().zip(keys_b.iter())
-            {
-                // SQLite treats NULL as smallest value:
-                // - ASC: NULL comes first (Less than non-NULL)
-                // - DESC: NULL comes last (Greater than non-NULL)
-                let cmp = match (val_a.is_null(), val_b.is_null()) {
-                    (true, true) => Ordering::Equal,
-                    (true, false) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Less
-                        } else {
-                            Ordering::Greater
-                        };
-                    }
-                    (false, true) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Greater
-                        } else {
-                            Ordering::Less
-                        };
-                    }
-                    (false, false) => match dir {
-                        vibesql_ast::OrderDirection::Asc => compare_sql_values(val_a, val_b),
-                        vibesql_ast::OrderDirection::Desc => {
-                            compare_sql_values(val_a, val_b).reverse()
-                        }
-                    },
-                };
-
-                if cmp != Ordering::Equal {
-                    return cmp;
-                }
-            }
-            Ordering::Equal
-        };
-
-        #[cfg(feature = "parallel")]
-        {
-            let config = ParallelConfig::global();
-            if config.should_parallelize_sort(rows.len()) {
-                rows.par_sort_by(comparison_fn);
-            } else {
-                rows.sort_by(comparison_fn);
-            }
-        }
-
-        #[cfg(not(feature = "parallel"))]
-        {
-            rows.sort_by(comparison_fn);
-        }
-
-        // Verify first few and last few are correctly sorted
-        assert_eq!(rows[0].0.values[0], SqlValue::Integer(0));
-        assert_eq!(rows[1].0.values[0], SqlValue::Integer(1));
-        assert_eq!(rows[2].0.values[0], SqlValue::Integer(2));
-        assert_eq!(rows[14997].0.values[0], SqlValue::Integer(14997));
-        assert_eq!(rows[14998].0.values[0], SqlValue::Integer(14998));
-        assert_eq!(rows[14999].0.values[0], SqlValue::Integer(14999));
-    }
-
-    #[test]
-    fn test_sort_descending_with_keys() {
-        let mut rows: Vec<RowWithSortKeys> = vec![
-            (
-                Row::from_vec(vec![SqlValue::Integer(1)]),
-                Some(vec![(SqlValue::Integer(1), vibesql_ast::OrderDirection::Desc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Integer(3)]),
-                Some(vec![(SqlValue::Integer(3), vibesql_ast::OrderDirection::Desc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Integer(2)]),
-                Some(vec![(SqlValue::Integer(2), vibesql_ast::OrderDirection::Desc, None, None)]),
-            ),
-        ];
-
-        let comparison_fn = |(_, keys_a): &RowWithSortKeys, (_, keys_b): &RowWithSortKeys| {
-            let keys_a = keys_a.as_ref().unwrap();
-            let keys_b = keys_b.as_ref().unwrap();
-
-            for ((val_a, dir, _nulls, _coll), (val_b, _, _, _)) in keys_a.iter().zip(keys_b.iter())
-            {
-                // SQLite treats NULL as smallest value:
-                // - ASC: NULL comes first (Less than non-NULL)
-                // - DESC: NULL comes last (Greater than non-NULL)
-                let cmp = match (val_a.is_null(), val_b.is_null()) {
-                    (true, true) => Ordering::Equal,
-                    (true, false) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Less
-                        } else {
-                            Ordering::Greater
-                        };
-                    }
-                    (false, true) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Greater
-                        } else {
-                            Ordering::Less
-                        };
-                    }
-                    (false, false) => match dir {
-                        vibesql_ast::OrderDirection::Asc => compare_sql_values(val_a, val_b),
-                        vibesql_ast::OrderDirection::Desc => {
-                            compare_sql_values(val_a, val_b).reverse()
-                        }
-                    },
-                };
-
-                if cmp != Ordering::Equal {
-                    return cmp;
-                }
-            }
-            Ordering::Equal
-        };
-
-        #[cfg(feature = "parallel")]
-        {
-            let config = ParallelConfig::global();
-            if config.should_parallelize_sort(rows.len()) {
-                rows.par_sort_by(comparison_fn);
-            } else {
-                rows.sort_by(comparison_fn);
-            }
-        }
-
-        #[cfg(not(feature = "parallel"))]
-        {
-            rows.sort_by(comparison_fn);
-        }
-
-        assert_eq!(rows[0].0.values[0], SqlValue::Integer(3));
-        assert_eq!(rows[1].0.values[0], SqlValue::Integer(2));
-        assert_eq!(rows[2].0.values[0], SqlValue::Integer(1));
-    }
-
-    #[test]
-    fn test_sort_with_nulls() {
-        // SQLite treats NULL as smallest value:
-        // - ASC: NULL comes first
-        // - DESC: NULL comes last
-        let mut rows_asc: Vec<RowWithSortKeys> = vec![
-            (
-                Row::from_vec(vec![SqlValue::Integer(2)]),
-                Some(vec![(SqlValue::Integer(2), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Null]),
-                Some(vec![(SqlValue::Null, vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-            (
-                Row::from_vec(vec![SqlValue::Integer(1)]),
-                Some(vec![(SqlValue::Integer(1), vibesql_ast::OrderDirection::Asc, None, None)]),
-            ),
-        ];
-
-        let comparison_fn = |(_, keys_a): &RowWithSortKeys, (_, keys_b): &RowWithSortKeys| {
-            let keys_a = keys_a.as_ref().unwrap();
-            let keys_b = keys_b.as_ref().unwrap();
-
-            for ((val_a, dir, _nulls, _coll), (val_b, _, _, _)) in keys_a.iter().zip(keys_b.iter())
-            {
-                // SQLite treats NULL as smallest value:
-                // - ASC: NULL comes first (Less than non-NULL)
-                // - DESC: NULL comes last (Greater than non-NULL)
-                let cmp = match (val_a.is_null(), val_b.is_null()) {
-                    (true, true) => Ordering::Equal,
-                    (true, false) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Less
-                        } else {
-                            Ordering::Greater
-                        };
-                    }
-                    (false, true) => {
-                        return if matches!(dir, vibesql_ast::OrderDirection::Asc) {
-                            Ordering::Greater
-                        } else {
-                            Ordering::Less
-                        };
-                    }
-                    (false, false) => match dir {
-                        vibesql_ast::OrderDirection::Asc => compare_sql_values(val_a, val_b),
-                        vibesql_ast::OrderDirection::Desc => {
-                            compare_sql_values(val_a, val_b).reverse()
-                        }
-                    },
-                };
-
-                if cmp != Ordering::Equal {
-                    return cmp;
-                }
-            }
-            Ordering::Equal
-        };
-
-        rows_asc.sort_by(&comparison_fn);
-
-        // ASC: NULL, 1, 2 (NULLs first, as smallest value)
-        assert_eq!(rows_asc[0].0.values[0], SqlValue::Null);
-        assert_eq!(rows_asc[1].0.values[0], SqlValue::Integer(1));
-        assert_eq!(rows_asc[2].0.values[0], SqlValue::Integer(2));
-    }
-
-    #[test]
-    #[cfg(feature = "parallel")]
-    fn test_parallel_config_threshold() {
-        let config = ParallelConfig::global();
-
-        // Verify that parallel sorting is disabled for small datasets
-        // (actual threshold depends on hardware, but should be > 100)
-        assert!(!config.should_parallelize_sort(100));
-
-        // Verify that parallel sorting is enabled for large datasets
-        // (15000 rows should trigger parallel path on any reasonable hardware)
-        // Note: On single-core systems this might still be false
-        let large_dataset_size = 15000;
-        let uses_parallel = config.should_parallelize_sort(large_dataset_size);
-
-        // Just verify the threshold logic is working (result depends on hardware)
-        if config.num_threads > 1 {
-            // On multi-core, large datasets should use parallel
-            assert!(uses_parallel || config.thresholds.sort > large_dataset_size);
-        }
     }
 }
